# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdminFeedbackPublic": {
        "description": "管理端视图：含提交环境上下文与处理标注。",
        "properties": {
          "admin_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Note"
          },
          "app_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "category": {
            "$ref": "#/components/schemas/FeedbackCategory"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "handled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handled At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "status": {
            "$ref": "#/components/schemas/FeedbackStatus"
          },
          "user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "user_id",
          "category",
          "content",
          "app_version",
          "platform",
          "status",
          "admin_note",
          "handled_at",
          "created_at"
        ],
        "title": "AdminFeedbackPublic",
        "type": "object"
      },
      "AdminFeedbackUpdateRequest": {
        "description": "管理端标注：改状态和/或写备注，至少动一项。",
        "properties": {
          "admin_note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Note"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AdminFeedbackUpdateRequest",
        "type": "object"
      },
...
      "AdminLetterStatusUpdate": {
        "description": "管理端信件状态流转。表内合法流转见 admin_service._ALLOWED_TRANSITIONS。",
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "$ref": "#/components/schemas/LetterStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "AdminLetterStatusUpdate",
        "type": "object"
      },
...
      "AdminLoginRequest": {
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "AdminLoginRequest",
        "type": "object"
      },
...
      "AdminTokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "title": "AdminTokenResponse",
        "type": "object"
      },
...
      "DeliveryMode": {
        "description": "投递方式（PRD 6.1：必选其一，两者都是一等公民）。",
        "enum": [
          "stay",
          "drift"
        ],
        "title": "DeliveryMode",
        "type": "string"
      },
...
      "FeedbackCategory": {
        "description": "反馈类型：问题反馈或改进建议。",
        "enum": [
          "bug",
          "suggestion"
        ],
        "title": "FeedbackCategory",
        "type": "string"
      },
...
      "FeedbackStatus": {
        "description": "反馈处理状态：待处理 → 已处理（管理端流转）。",
        "enum": [
          "open",
          "resolved"
        ],
        "title": "FeedbackStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LetterStatus": {
        "description": "信件状态（见 docs/ARCHITECTURE.md §3 状态机）。\n\n只有 PUBLIC 参与漂流抽取与就地发掘。\n机审失效时默认停在 PENDING —— 降级方向永远更保守（PRD §8.2）。",
        "enum": [
          "pending",
          "public",
          "rejected",
          "taken_down"
        ],
        "title": "LetterStatus",
        "type": "string"
      },
...
      "Page_AdminFeedbackPublic_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminFeedbackPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[AdminFeedbackPublic]",
        "type": "object"
      },
      "Page_AdminLetterSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminLetterSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[AdminLetterSummary]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "匿名漂流信系统。信可随机漂向远方（drift），也可埋在某地等后来者发掘（stay）。\n\n**匿名铁律**：一切信件响应使用 LetterPublic，不含任何作者标识。",
    "title": "风信 Kaze no tayori API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/feedbacks": {
      "get": {
        "description": "反馈列表，created_at DESC，可按状态/类型筛选。",
        "operationId": "list_feedbacks_v1_admin_feedbacks_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FeedbackStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FeedbackCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AdminFeedbackPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Feedbacks",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/feedbacks/{feedback_id}": {
      "patch": {
        "description": "标注反馈：改状态（open↔resolved，联动 handled_at）和/或写备注。",
        "operationId": "update_feedback_v1_admin_feedbacks__feedback_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "feedback_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Feedback Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminFeedbackUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFeedbackPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Feedback",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/letters": {
      "get": {
        "description": "信件列表（含非 public），created_at DESC。owner=seed|user 区分种子信/有主信。",
        "operationId": "list_letters_v1_admin_letters_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LetterStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "delivery_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DeliveryMode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Mode"
            }
          },
          {
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(seed|user)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AdminLetterSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Letters",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/letters/{letter_id}": {
      "get": {
        "description": "信件详情（全状态可见 + owner_user_id，仅管理端可见）。",
        "operationId": "get_letter_v1_admin_letters__letter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "letter_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Letter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLetterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Letter",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/letters/{letter_id}/status": {
      "patch": {
        "description": "审核/下架/恢复/赦免。状态机见 admin_service._ALLOWED_TRANSITIONS，表外 409。",
        "operationId": "transition_letter_status_v1_admin_letters__letter_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "letter_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Letter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLetterStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLetterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Transition Letter Status",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/login": {
      "post": {
        "description": "用户名密码换 admin token（12h 短时效）。凭据错误一律 401。",
        "operationId": "admin_login_v1_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Login",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/reports": {
      "get": {
        "description": "举报列表（JOIN 涉事信摘要），created_at DESC。默认只看 open。",
        "operationId": "list_reports_v1_admin_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReportStatus",
              "default": "open"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AdminReportPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Reports",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/reports/{report_id}": {
      "patch": {
        "description": "举报处置：dismissed / actioned（置已处理回写 handled_at，回退 open 清空）。",
        "operationId": "update_report_v1_admin_reports__report_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReportUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Report",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/seed-letters": {
      "get": {
        "description": "种子信列表（owner IS NULL，含被下架的），created_at DESC。",
        "operationId": "list_seed_letters_v1_admin_seed_letters_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AdminLetterSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Seed Letters",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "新建种子信：复用写信校验，owner=NULL、直接 public 入池（跳过机审）。",
        "operationId": "create_seed_letter_v1_admin_seed_letters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LetterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLetterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Seed Letter",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/seed-letters/{letter_id}": {
      "patch": {
        "description": "编辑种子信：blocks/文案/落点/天气可改；theme 绑定不给改；仅限无主信。",
        "operationId": "update_seed_letter_v1_admin_seed_letters__letter_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "letter_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Letter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSeedLetterUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLetterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Seed Letter",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/stats": {
      "get": {
        "description": "概览聚合：状态分布 / 用户数 / 7·30 日新增 / 池健康 / 待办角标。",
        "operationId": "get_stats_v1_admin_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStats"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Stats",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/letters": {
      "post": {
        "description": "写一封信并投放。\n\ndelivery_mode 必选：stay（锚定位置）或 drift（入随机漂流池）。\n提交后入审，默认 status=pending。\n审核通过的公开信在响应后由后台任务自动补写 AI 短诗\n（客户端已自带 poem 或 FEATURE_AI 关闭时跳过）。",
        "operationId": "create_letter_v1_letters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LetterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LetterOwned"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Letter",
        "tags": [
          "letters"
        ]
      }
    },
    "/v1/letters/{letter_id}": {
      "get": {
        "description": "读单封公开信（回信溯源用）。非 public 一律 404。\n\n纯读：不写 letter_reads、不动 read_count（开信上报走 POST /letters/{id}/read）。\n登录读者附带 me_resonated——点过共鸣的章常亮，重进未亮是反直觉的。",
        "operationId": "read_letter_v1_letters__letter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "letter_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Letter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LetterPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Letter",
        "tags": [
          "letters"
        ]
      }
    },
...
  }
}